Remove unecessary circuit.copy() call from construct_evaluation_circuit
Right now in the construct_evaluation_circuit() in the
WeightedPauliOperator class we're needlessly copying a circuit twice.
The copy() method of QuantumCircuit objects is expensive because it ends
up running deepcopy() on each Instruction object, which ends up becoming
a bottleneck for very deep circuits. This commit partially mitigates
this by removing the first outer copy and just constructing a new
circuit from the waveform function circuit copy for each basis.

There are still several instances of QuantumCircuit.copy() in the call
path which still mean for larger circuits deepcopy remains the
bottleneck for execution speed. But, removing that copy usage will
require further investigation to see if it can be accomplished without
changing the interface to the methods.